Add support for multiple compilation units (fix #164)
This is a breaking change (see the PR notes)

 Please enter the commit message for your changes. Lines starting
