+++ crytic_compile/platform/hardhat.py
@@ -17,6 +17,7 @@ from .abstract_platform import AbstractPlatform
 
 # Handle cycle
 from .solc import relative_to_short
+from ..compilation_unit import CompilationUnit
 
 if TYPE_CHECKING:
     from crytic_compile import CryticCompile
@@ -45,9 +46,6 @@ class Hardhat(AbstractPlatform):
         hardhat_ignore_compile = kwargs.get("hardhat_ignore_compile", False) or kwargs.get(
             "ignore_compile", False
         )
-
-        cache_directory = kwargs.get("hardhat_cache_directory", "cache")
-        config_file = Path(cache_directory, "solidity-files-cache.json")
 
         build_directory = Path(kwargs.get("hardhat_artifacts_directory", "artifacts"), "build-info")
 
@@ -79,75 +77,88 @@ class Hardhat(AbstractPlatform):
             if stderr:
                 LOGGER.error(stderr)
 
-        (compiler, version_from_config, optimized) = _get_version_from_config(config_file)
-
-        crytic_compile.compiler_version = CompilerVersion(
-            compiler=compiler, version=version_from_config, optimized=optimized
-        )
-
-        skip_filename = crytic_compile.compiler_version.version in [
-            f"0.4.{x}" for x in range(0, 10)
-        ]
-
         files = sorted(
             os.listdir(build_directory), key=lambda x: os.path.getmtime(Path(build_directory, x))
         )
+        files = [f for f in files if f.endswith(".json")]
         if not files:
             txt = f"`hardhat compile` failed. Can you run it?\n{build_directory} is empty"
             raise InvalidCompilation(txt)
 
-        build_info = Path(build_directory, files[0])
-        with open(build_info, encoding="utf8") as file_desc:
-            targets_json = json.load(file_desc)["output"]
+        for file in files:
+            build_info = Path(build_directory, file)
 
-            if "contracts" in targets_json:
-                for original_filename, contracts_info in targets_json["contracts"].items():
-                    for original_contract_name, info in contracts_info.items():
-                        contract_name = extract_name(original_contract_name)
+            compilation_unit = CompilationUnit(crytic_compile, file)
 
-                        contract_filename = convert_filename(
-                            original_filename,
-                            relative_to_short,
-                            crytic_compile,
-                            working_dir=hardhat_working_dir,
-                        )
+            with open(build_info, encoding="utf8") as file_desc:
+                loaded_json = json.load(file_desc)
 
-                        crytic_compile.contracts_names.add(contract_name)
-                        crytic_compile.contracts_filenames[contract_name] = contract_filename
+                targets_json = loaded_json["output"]
 
-                        crytic_compile.abis[contract_name] = info["abi"]
-                        crytic_compile.bytecodes_init[contract_name] = info["evm"]["bytecode"][
-                            "object"
-                        ]
-                        crytic_compile.bytecodes_runtime[contract_name] = info["evm"][
-                            "deployedBytecode"
-                        ]["object"]
-                        crytic_compile.srcmaps_init[contract_name] = info["evm"]["bytecode"][
-                            "sourceMap"
-                        ].split(";")
-                        crytic_compile.srcmaps_runtime[contract_name] = info["evm"][
-                            "deployedBytecode"
-                        ]["sourceMap"].split(";")
-                        userdoc = info.get("userdoc", {})
-                        devdoc = info.get("devdoc", {})
-                        natspec = Natspec(userdoc, devdoc)
-                        crytic_compile.natspec[contract_name] = natspec
+                version_from_config = loaded_json["solcVersion"]  # TODO supper vyper
+                input_json = loaded_json["input"]
+                compiler = "solc" if input_json["language"] == "Solidity" else "vyper"
+                optimized = input_json["settings"]["optimizer"]["enabled"]
 
-            if "sources" in targets_json:
-                for path, info in targets_json["sources"].items():
-                    if skip_filename:
-                        path = convert_filename(
-                            self._target,
-                            relative_to_short,
-                            crytic_compile,
-                            working_dir=hardhat_working_dir,
-                        )
-                    else:
-                        path = convert_filename(
-                            path, relative_to_short, crytic_compile, working_dir=hardhat_working_dir
-                        )
-                    crytic_compile.filenames.add(path)
-                    crytic_compile.asts[path.absolute] = info["ast"]
+                compilation_unit.compiler_version = CompilerVersion(
+                    compiler=compiler, version=version_from_config, optimized=optimized
+                )
+
+                skip_filename = compilation_unit.compiler_version.version in [
+                    f"0.4.{x}" for x in range(0, 10)
+                ]
+
+                if "contracts" in targets_json:
+                    for original_filename, contracts_info in targets_json["contracts"].items():
+                        for original_contract_name, info in contracts_info.items():
+                            contract_name = extract_name(original_contract_name)
+
+                            contract_filename = convert_filename(
+                                original_filename,
+                                relative_to_short,
+                                crytic_compile,
+                                working_dir=hardhat_working_dir,
+                            )
+
+                            compilation_unit.contracts_names.add(contract_name)
+                            compilation_unit.contracts_filenames[contract_name] = contract_filename
+
+                            compilation_unit.abis[contract_name] = info["abi"]
+                            compilation_unit.bytecodes_init[contract_name] = info["evm"][
+                                "bytecode"
+                            ]["object"]
+                            compilation_unit.bytecodes_runtime[contract_name] = info["evm"][
+                                "deployedBytecode"
+                            ]["object"]
+                            compilation_unit.srcmaps_init[contract_name] = info["evm"]["bytecode"][
+                                "sourceMap"
+                            ].split(";")
+                            compilation_unit.srcmaps_runtime[contract_name] = info["evm"][
+                                "deployedBytecode"
+                            ]["sourceMap"].split(";")
+                            userdoc = info.get("userdoc", {})
+                            devdoc = info.get("devdoc", {})
+                            natspec = Natspec(userdoc, devdoc)
+                            compilation_unit.natspec[contract_name] = natspec
+
+                if "sources" in targets_json:
+                    for path, info in targets_json["sources"].items():
+                        if skip_filename:
+                            path = convert_filename(
+                                self._target,
+                                relative_to_short,
+                                crytic_compile,
+                                working_dir=hardhat_working_dir,
+                            )
+                        else:
+                            path = convert_filename(
+                                path,
+                                relative_to_short,
+                                crytic_compile,
+                                working_dir=hardhat_working_dir,
+                            )
+                        crytic_compile.filenames.add(path)
+                        compilation_unit.asts[path.absolute] = info["ast"]
 
     @staticmethod
     def is_supported(target: str, **kwargs: str) -> bool:
